add wallet/types.rs
